Cleaned up Observable components for master branch integration
