+++ gatsby-browser.js
@@ -5,6 +5,3 @@
  */
 
 // You can delete this file if you're not using it
-
-import reduxStore from "./redux-store"
-export const wrapRootElement = reduxStore
