+++ src/components/observable-runtime.tsx
@@ -1,8 +1,7 @@
 import React from "react";
 import {Runtime, Inspector} from "@observablehq/notebook-runtime";
 import notebook from "windfall-awareness-notebook-prototype";
-import { node } from "prop-types";
-import { generateRuntime } from './observable-dock';
+import { generateRuntime } from './generate-runtime';
 
 export const ObservableContext = React.createContext();
 
@@ -10,39 +9,22 @@ export const ObservableContext = React.createContext();
 export class ObservableRuntime extends React.Component {
   constructor(props, context) {
       super(props, context);
-      this.registerCellRef = this.registerCellRef.bind(this);
-      this.resetCellRefs = this.resetCellRefs.bind(this);
+      
       this.storeCellValues = this.storeCellValues.bind(this);
-      this.generateDom = this.generateDom.bind(this);
       this.resetHTML = this.resetHTML.bind(this);
-      this.premain = () => {if (typeof window !== 'undefined') {console.log(window); return generateRuntime(notebook)}}
+      
+      //Generate Observable Notebook runtime object, this contains module namespace
+      //premain() to prevent build error reference to window. 
+      this.premain = () => {if (typeof window !== 'undefined') {return generateRuntime(notebook)}}
       this.main = this.premain()
       
       this.observer = (dom) => {return new Inspector(dom)}
       
       this.state = {
-        runtime: '',
-        childrenCellRefs: [],
-        runtimeState: {},
-        allCells: [],
-        allNamedCells: [],
         valueStore: {},
         htmlList: {}
       };
    }
-
-  generateDom(cell) {
-    this.state.allCells.push(cell)
-    this.main.variable(this.observer(cell.domRef)).define([cell.cellName], widget => widget)
-  }
-
-  registerCellRef(cellName, domRef) {
-    this.state.childrenCellRefs.push({cellName, domRef})
-  }
-
-  resetCellRefs() {
-    this.state.childrenCellRefs = []
-  }
 
   storeCellValues(cellName, cellValue) {
       
@@ -65,8 +47,8 @@ export class ObservableRuntime extends React.Component {
     }
   }
 
+  //Reset Observable Cell nodes to allow Observable Inspector to reinsert div.
   resetHTML(cellName, cellNode) {
-    
     this.storeCellValues(cellName, cellNode)
     if (cellNode) {
         cellNode.innerHTML = this.state.htmlList[cellName]
@@ -75,10 +57,7 @@ export class ObservableRuntime extends React.Component {
 
   render() {
       return <ObservableContext.Provider value={{
-           registerCellRef: this.registerCellRef, 
-           resetCellRefs: this.resetCellRefs,
-           runtime: this.state.runtime,
-           testmain: this.main,
+           main: this.main,
            observer: this.observer,
            resetHTML: this.resetHTML
            }}>
